Added some more information management calls

+++ template.c
@@ -12,19 +12,19 @@
 #include <sys/ioctl.h>
 #include <stdlib.h>
 #include <fcntl.h>
+#include <sys/vfs.h>
 #include <ifaddrs.h>
 #include <linux/random.h>
-
 
 int main(int argc, char* argv[])
 {
 	struct timeval start;
 	struct timeval end;
 	long diff;
-	int rc;
-	// Extra Code
+        long rc;
+        // Extra Code
 	gettimeofday(&start, NULL);
-	rc = XXX;
+	rc = (long)XXX;
 	gettimeofday(&end, NULL);
 	if(rc == -1)
 		printf("-1,");
@@ -32,3 +32,5 @@ int main(int argc, char* argv[])
 		diff = (end.tv_sec - start.tv_sec) * 1000000 + (end.tv_usec - start.tv_usec);
 	printf("%li,", diff);
 }
+
+// https://www.improgrammer.net/type-casting-c-language/
